// frontend/src/api/public/model/components-schemas-todo/todoResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Good Todo API - Public
 * Public API for Good Todo application
 * OpenAPI spec version: 0.0.1
 */

export interface TodoResponse {
  id?: string;
  title?: string;
  description?: string;
  completed?: boolean;
  /** If true, visible to all users in the same tenant */
  is_public?: boolean;
  /** @nullable */
  due_date?: string | null;
  /** @nullable */
  completed_at?: string | null;
  created_at?: string;
  updated_at?: string;
}